+++ db.json
@@ -1,7 +1,7 @@
 {
   "books": [
     {
-      "title": "Harry Potter.",
+      "title": "Harry Potter",
       "id": 19235
     },
     {
