+++ src/components/BookEdit.tsx
@@ -1,12 +1,15 @@
 import { PlusCircleIcon } from '@heroicons/react/24/outline';
 import style from './BookShow.module.css';
-import { useState } from 'react';
+import { useContext, useState } from 'react';
 import { Book } from '../types';
+import { BookContext } from '../context/books';
 type props = {
   book: Book;
   handleSubmit: (book: Book) => void;
 };
 export default function BookEdit({ book, handleSubmit }: props) {
+  const ctx = useContext(BookContext);
+  console.log(ctx);
   const [editTitle, setEditTitle] = useState(book.title);
 
   const onChangeInputHandler = (event: React.ChangeEvent<HTMLInputElement>) => {
@@ -23,7 +26,13 @@ export default function BookEdit({ book, handleSubmit }: props) {
       }}
     >
       <label htmlFor='title'>
-        <input type='text' name='title' value={editTitle} onChange={onChangeInputHandler} />
+        <input
+          type='text'
+          name='title'
+          id='title'
+          value={editTitle}
+          onChange={onChangeInputHandler}
+        />
       </label>
       <button className={style.card__btn} type='submit'>
         <PlusCircleIcon width={25} />
